requirements.txt
Flask>=1.0.0
cloudant==2.13.0
discord.py==1.4.1
rivescript==1.14.9
pyowm==3.0.0
python_dotenv==0.14.0
requests==2.24.0
wikipedia==1.4.0
google==2.0.3
json5==0.9.4
pynacl==1.4.0
pytube4==0.1.1
youtube-search==1.1.0
